+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `defaultValue` to `files` form element validation
+- `excludedElementIds` to `pdf.generateFormSubmissionPDF()`
 
 ## [0.9.15] - 2021-06-23